Change float32 to 64 bit for improve precision Fix minor syntax error

+++ database.go
@@ -52,7 +52,7 @@ func getCustomer(custID int) Customer {
 	return customer
 }
 
-func setCustomerBalance(custID int, balance float32) bool {
+func setCustomerBalance(custID int, balance float64) bool {
 	customer := getCustomer(custID)
 	if customer.ID == -1 {
 		return false
