+++ main.go
@@ -3,7 +3,7 @@ package main
 //Customer structure
 type Customer struct {
 	ID      int     `json:"id"`
-	Balance float32 `json:"balance"`
+	Balance float64 `json:"balance"`
 }
 
 func main() {
